Fix physics scale
Physics scale from pixels to Box2D units was 1:1, Box2D actively recomends against this.
https://box2d.org/posts/2011/12/pixels/
We are now using meters as units (player character is 1.5 units tall and 1 unit wide), and scaling up during rendering.
